requirements.txt
## Base 
matplotlib>=3.2.2
numpy>=1.18.5,<1.24.0
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
imgaug
logbook
dvc
dvc-gdrive

## yolov7 & yolov5 polygon
torch>=1.7.0,!=1.12.0
torchvision>=0.8.1,!=0.13.0
tqdm>=4.41.0
protobuf<4.21.3
pandas>=1.1.4
seaborn>=0.11.0
shapely
pycocotools
onnx==1.13.1
tensorboard

## vietocr
einops==0.2.0

## api
fastapi==0.94.1
uvicorn==0.21.1
streamlit==1.20.0
pydantic==1.10.6
pyngrok
